packages/stream/pyproject.toml
[project]
name = "flux0-stream"
version = "0.0.1"
requires-python = ">=3.12"
dependencies = [
    "flux0-core",
    "jsonpatch>=1.33",
]

[project.optional-dependencies]
langchain = [
    "langchain-core>=0.3.39",
]

[tool.uv.sources]
flux0-core = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "deepdiff>=8.2.0",
    "pytest-asyncio>=0.25.3",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
pythonpath = [
  "."
]
